move thunks to own file, move all day and multi day actions to thunks
